Add custom SQLite3 PHP Session Handler
PHP allows you to define custom handlers for its sessions, instead of using the
default one that relies on storing files. This is done by implementing the
SessionHandlerInterface interface and then calling session_set_save_handler().

This custom session handler stores all session data in an SQLite3 database,
with the idea being that we can use the same database as the rendezvous
application data. Storing this in a single database allows for full control of
the information by the application, and all state is in a single file. There's
no other dependency, and portability is enhanced.

Normally the SessionHandlerInterface interface is designed to be run centrally,
in the PHP server, and was heavily inspired by the file-per-session defaults so
it makes some assumptions that are not true in our case. For this reason, the
particular implementation decides to deviate from expectations (as evident in
the open() and close() calls) and instead works on a per-page basis with every
request reopening (and closing) the database from scratch. Particular attention
has been paid to the permissions used each time, to ensure more efficient
locking and unlocking by SQLite3.

Finally, this is the first piece of code in the rendezvous project that has
tests! It felt easy to write them, and I caught a lot of issues that would not
have been visible otherwise. There's no standard framework, nor are they unit
tests. Instead, it's a single PHP file, that is meant to be executed via the
CLI. It runs a scenario and prints the first test that failed, and returns 1 if
there are issues. If all tests pass, it returns 0 and does not print anything.
It is not so flexible and stateless, and a lot of the tests rely on previous
test results, but it can reasonably detect problems and common situations that
it might find itself in.