Replace txtDB with SQLite3 in index.php
This commit replaces all use of SQLite3 with txtDB in the `index.php` file.
From now on, a new model is being followed, where all database operations are
being outsourced to `db.php` and the main code only includes high-level
function calls. All the complexity is within this file, and only a simple error
check is required by the callee.